Fix create processing

+++ src/main/java/priam/data/priamdataservice/entities/Consent.java
@@ -13,7 +13,7 @@ import javax.persistence.*;
 @Table(name="consent")
 public class Consent {
     @Id
-    private int dataSubjectId;
+    private Long dataSubjectId;
     @Id
-    private int processingId;
+    private Long processingId;
 }
