+++ src/main/java/priam/data/priamdataservice/entities/ConsentKey.java
@@ -3,6 +3,6 @@ package priam.data.priamdataservice.entities;
 import java.io.Serializable;
 
 public class ConsentKey implements Serializable {
-    private int dataSubjectId;
-    private int processingId;
+    private Long dataSubjectId;
+    private Long processingId;
 }
