+++ src/main/java/priam/data/priamdataservice/entities/ProcessedDataKey.java
@@ -3,6 +3,6 @@ package priam.data.priamdataservice.entities;
 import java.io.Serializable;
 
 public class ProcessedDataKey implements Serializable {
-    private int dataSubjectId;
-    private int dataId;
+    private Long dataSubjectId;
+    private Long dataId;
 }
